[wasm][liftoff] Don't reserve space for the stack marker twice
The total frame size returned by GetTotalFrameSize includes the frame
marker. However, the frame marker is pushed on the stack with a push
instruction. Therefore it is not needed to allocate memory for it again
on the stack. This CL therefore reduces the memory allocated on the
stack by the size of the frame marker.

R=clemensb@chromium.org

Bug: v8:11074
Change-Id: Ie04508a57a2c641a2ee5d89d72dd22ec0572b5e5
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/2557510
Reviewed-by: Clemens Backes <clemensb@chromium.org>
Commit-Queue: Andreas Haas <ahaas@chromium.org>
Cr-Commit-Position: refs/heads/master@{#71373}