[wasm] Use BrOrRet in Interface for conditional branches
The logic to detect if a 'br' instruction is a branch or a return was
duplicated in function-body-decoder-impl.h and in both interfaces.
Apart from code duplication, this structure also made it hard to
implement planned compiler improvements.
This CL removes this duplication by upgrading BrOrRet (that already
existed in both Liftoff and Turbofan interfaces) to an interface
function and using it for unconditional branches.

Change-Id: Ia04952cce621335268fc40ef9544a99b61dc7da3
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/2557515
Commit-Queue: Manos Koukoutos <manoskouk@chromium.org>
Reviewed-by: Clemens Backes <clemensb@chromium.org>
Cr-Commit-Position: refs/heads/master@{#71400}